broken stick nav. still working
