+++ Gemfile
@@ -1,6 +1,7 @@
 source 'https://rubygems.org'
 
-
+gem 'responders', '~> 2.0'
+gem "font-awesome-rails"
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '4.2.8'
 # Use postgres as the database for Active Record
