+++ app/assets/stylesheets/_variables.scss
@@ -1,0 +1,8 @@
+
+
+/***************** Offsets **********************/
+
+$banner-top-padding-lg: 25px;
+$banner-top-padding-sm: 10px;
+$banner-text-sm: 46px;
+$banner-text-lg: 96px;
