--- app/assets/stylesheets/custom.scss
@@ -1,59 +1,0 @@
- /*Move down content because we have a fixed navbar that is 50px tall */
-
-
-.instacart-green {
-	background-color: #68BF54!important;
-}
-
-.text_color {
-	color: #fff
-}
-
-.form-group {
-  padding: 10px;
-  align-content: center;
-}
-
-.display-3 {
-	font-family: 'Muli', sans-serif;
-/*  font-family: 'Montserrat', sans-serif;		
-*/}
-
-footer{
-  /*background-color: white; */
-}
-
-
-
-body,
-.landing-page {
-  .overlay{
-    background-color: rgba(0,0,0,0.3);
-    background-size: cover;
-  }
-
-  height: 100%;
-  padding-bottom: 0px;
-
-  font-family: 'Muli', sans-serif;
-  /*font-family: 'Montserrat', sans-serif;*/
-
-  .jumbotron{
-      margin-top: auto;
-      /*background: url('../img/carrots.jpg');*/
-      background-color: rgba(0,0,0,0.0);
-      background-size: cover;
-      /*opacity: .9;*/
-        /*filter: alpha(opacity=50);*/
-        text-shadow: 1.5px 1px 0 black;
-      padding-bottom: 1em;
-    }
-}
-
-.landing-page {
-    background: url("carrots.jpg")no-repeat center center fixed;
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-}
