+++ app/assets/stylesheets/landing_page.scss
@@ -43,21 +43,18 @@ body,
   /*font-family: 'Montserrat', sans-serif;*/
 
   .jumbotron{
-      margin-top: auto;
-      /*background: url('../img/carrots.jpg');*/
+      margin-top: 50px;
+      // margin-top: auto;
       background-color: rgba(0,0,0,0.0);
       background-size: cover;
       /*opacity: .9;*/
         /*filter: alpha(opacity=50);*/
-        text-shadow: 1.5px 1px 0 black;
+      text-shadow: 1.5px 1px 0 black;
       padding-bottom: 1em;
     }
 }
 
 .landing-page {
     background: url("carrots.jpg")no-repeat center center fixed;
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
     background-size: cover;
 }
