--- app/assets/stylesheets/question.scss
@@ -1,51 +1,0 @@
-// Place all the styles related to the Question controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
-
-
-
-
-.question-modal{
-	.modal-content, .modal-body{
-		border-radius: 0px;
-		border: none;
-	}
-
-	.modal-header{
-		padding-right: 10px;
-		padding-left: 10px;
-		padding-top: 10px;
-		padding-bottom: 5px;
-		border: none;
-	}
-
-	form{
-		textarea{
-			border-radius: 0px;
-			background-color: white;		
-			border-right: none;
-			border-top: none;	
-			line-height: 30%;
-			transition: all 0.2s linear;
-
-			&:focus{
-				background-color: rgba(223, 255, 175, 0.2);
-				border-left: none;
-				line-height: 100%;
-				transition: all 0.2s linear;
-			}
-			&:hover{
-				box-shadow: 0px 4px 5px -6px rgba(0,0,0,0.75);
-			}
-		}
-	}
-	label{
-		border-bottom: 0.5px solid rgba(220, 220, 220, 0.7);
-		color: rgb(15, 119, 26);
-		box-shadow: 0px 4px 5px -6px rgba(0,0,0,0.75);
-	}
-	input{
-		border-radius: 0px;
-		border-top: none;
-	}
-}
