+++ app/assets/stylesheets/questions.scss
@@ -1,0 +1,208 @@
+// Place all the styles related to the Question controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
+@import 'variables'; // colors, fonts etc...
+
+
+ /******************* GENERAL **********************************/
+// body{
+// 	padding: 0px;
+// }
+
+
+
+ /******************* ADD QUESTION MODAL **********************************/
+
+.question-modal{
+	.modal-content, .modal-body{
+		border-radius: 0px;
+		border: none;
+	}
+
+	.modal-header{
+		padding-right: 10px;
+		padding-left: 10px;
+		padding-top: 10px;
+		padding-bottom: 5px;
+		border: none;
+	}
+
+	form{
+		textarea{
+			border-radius: 0px;
+			background-color: white;		
+			border-right: none;
+			border-top: none;	
+			line-height: 30%;
+			transition: all 0.2s linear;
+
+			&:focus{
+				background-color: rgba(223, 255, 175, 0.2);
+				border-left: none;
+				line-height: 100%;
+				transition: all 0.2s linear;
+			}
+			&:hover{
+				box-shadow: 0px 4px 5px -6px rgba(0,0,0,0.75);
+			}
+		}
+	}
+	label{
+		border-bottom: 0.5px solid rgba(220, 220, 220, 0.7);
+		color: rgb(15, 119, 26);
+		box-shadow: 0px 4px 5px -6px rgba(0,0,0,0.75);
+	}
+	input{
+		border-radius: 0px;
+		border-top: none;
+	}
+}
+
+/******************* Questions Index Navbar **********************************/
+
+
+body.fixed-nav nav {
+  position: fixed;
+  top: 0;
+  box-shadow:0 5px 0 rgba(0,0,0,0.1);
+  width: 100%
+}
+
+.navbar{
+	background-color: #68BF54; // #008ba3; //#087f23;
+	width:100%;
+	position: relative;
+	box-shadow: 2px -5px 24px -6px rgba(0,0,0,0.75);
+
+	.hidden-nav {
+	  opacity: 0;
+	  overflow: hidden;
+	  max-width:0;
+	  padding: 0;
+	  background: white;
+	  transition: ease-out 0.4s;
+	}
+
+	.hide-nav {
+	  opacity: 1;
+	  max-width:500px;
+	  padding-right: 2rem;
+	  transition: ease-out 0.4s;
+	}
+
+	.fixed-nav a.hide-nav  {
+	  opacity: 0;
+	  overflow: hidden;
+	  max-width:0;
+	  padding: 0;
+	  background: white;
+	  transition: ease-out 0.4s;
+	  font-size: 1px
+
+	}
+
+
+	.navbar-logo{
+		background-color: white;
+		padding: 3px;
+		border-radius: 5px;
+		border-right: 0.3px solid rgba(20,50,40,.5);
+		border-bottom: 1px solid rgba(20,50,40,.5);
+		box-shadow: 2px 2px 4px -2px rgba(20,50,40,.5);
+		&:hover{
+			box-shadow: 4px 4px 4px -2px rgba(20,50,40,.5);
+		}
+	}
+
+	input {
+		outline: none;
+	}
+	input[type=search]{
+		-webkit-appearance: textfield;
+		-webkit-box-sizing: content-box;
+		font-family: inherit;
+		font-size: 100%;
+	}
+
+
+	input[type=search]{
+		background: #ededed url(https://static.tumblr.com/ftv85bp/MIXmud4tx/search-icon.png) no-repeat 9px center;
+		border: solid 1px #ccc;
+		padding: 9px 10px 9px 32px;
+		width: 100px;
+				
+		transition:  0.7s cubic-bezier(0.61,-0.19, 0.7,-0.11),
+	}
+	input[type=search]:focus{
+		width: 350px;
+		background-color: #fff;
+		border-color: #66CC75;
+		box-shadow: 0 0 5px rgba(109,207,246,.5);
+	}
+
+	input[type=search]:hover{
+		background-color: #fff;
+	}
+
+}
+
+/******************* Questions Index Banner **********************************/
+
+.banner{
+	background: url("carrots.jpg")no-repeat center center fixed;
+  background-size: cover;
+  // background-color: #4caf50;
+  // background-color: #00acc1;
+  // opacity: 1;
+  // filter:blur(1px);
+
+
+  .overlay{
+  	padding-bottom: 1em;
+  	padding-left: 2em;
+		padding-top: $banner-top-padding-sm;
+		background-color: rgba(0,0,0,0.3);
+  // background-color: white;
+
+  }
+
+
+  .page-title{
+  	display: -webkit-flex;
+  	display: flex;
+  	-webkit-flex-direction: row; /* works with row or column */
+  	flex-direction: row;
+  	-webkit-align-items: center;
+  	align-items: center;
+  	-webkit-justify-content: center;
+  	justify-content: center;
+  }
+
+  .banner-image{
+  	flex: 1;
+  	width: 60px;
+  	height: 60px;
+  }
+
+  .banner-text{
+  		flex: 8;
+  		font-family: 'Muli', sans-serif;
+  	  font-weight: 300;
+  		// color: rgb(15, 79, 26);
+  		color: white;
+  	  text-shadow: 1.5px 1.5px 0 black; //rgba(20, 20, 10, 0.7);
+  	  margin-bottom: 0px;
+  	  font-size: $banner-text-sm;
+  }
+
+  @media(min-width: 768px){
+  	.banner-text{
+	  	font-size: $banner-text-lg;
+  	}
+  	.banner-image{
+  		width: 80px;
+  		height: 80px;
+  	}
+  }
+
+}
