+++ app/models/answer.rb
@@ -1,0 +1,2 @@
+class Answer < ActiveRecord::Base
+end
