+++ app/views/landing_page/index.html.erb
@@ -1,39 +1,9 @@
-
-  <head>
-
-
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css" integrity="sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ" crossorigin="anonymous">
-  
-  <!-- Google Font -->
-    <link href='http://fonts.googleapis.com/css?family=Oswald:400,300,700' rel='stylesheet' type='text/css'>
-
-    <link href="https://fonts.googleapis.com/css?family=Montserrat" rel="stylesheet">
-
-    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet">
-  <!-- Google Font -->
-
-
-   <script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-      <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/js/bootstrap.min.js" integrity="sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn" crossorigin="anonymous"></script>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="">
-    <meta name="author" content="">
-  
-
-    <title>InstaSmart</title>
-
-    <!-- Bootstrap core CSS -->
-
-
-  </head>
-    <body class="landing-page">
+    <div class="landing-page">
       <div class="overlay">
 
       <!-- NAVBAR -->
 
-      <%= render 'navbar' %>
+      <%= render 'landing_navbar' %>
 
       <!-- JUMBOTRON-->
       <div class="jumbotron text_color">
@@ -52,5 +22,5 @@
       <footer>
         <p>&copy; Instacart 2017</p>
       </footer>
-   </body>
+   </div>
