+++ app/views/layouts/application.html.erb
@@ -12,17 +12,18 @@
   <script src="https://code.jquery.com/jquery-3.1.1.slim.min.js" integrity="sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n" crossorigin="anonymous"></script>
   <script>window.jQuery || document.write('<script src="../../assets/js/vendor/jquery.min.js"><\/script>')</script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-
   <!-- Custom styles for this template -->
-  <!-- <%#= stylesheet_link_tag "custom" %> -->
+  <%= stylesheet_link_tag "_variables" %> 
   <link rel="icon" href="favicon.ico">
+  <link href="https://fonts.googleapis.com/css?family=Crimson+Text|Gentium+Basic" rel="stylesheet">
+  <link href="https://fonts.googleapis.com/css?family=Boogaloo" rel="stylesheet">
 
-
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css" integrity="sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ" crossorigin="anonymous">
 </head>
 <body>
-
-<%= yield %>
-  <%= render "question/question_modal", question: Question.new %>   
+  <%= render "questions/question_modal", question: Question.new %>   
+  <%= yield %>
+  
 
 </body>
 </html>
