+++ app/views/questions/_banner.html.erb
@@ -1,0 +1,10 @@
+	
+	<div class="banner" >
+		<div class="overlay">
+
+			<div class="page-title">
+	      <%= image_tag("carrot_2.png", :class=>"banner-image", :alt => "logo") %>
+				<p class="banner-text"><i>Questions</i></p>
+			</div>
+		</div>
+	</div>
