+++ app/views/questions/_navbar.html.erb
@@ -1,0 +1,31 @@
+
+<nav class="navbar navbar-toggleable-md navbar-inverse">
+     <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarsExampleDefault" aria-controls="navbarsExampleDefault" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <a class="navbar-brand hidden-nav" href="index.html">
+      <!-- <img src="/images/logo.png" width="120" height="30" alt="logo"> -->
+      <%= image_tag("carrot_2.png", :class=>"navbar-logo", :alt => "logo", :width => '40', :height => '40') %>
+    </a>
+
+    <div class="collapse navbar-collapse" id="navbarsExampleDefault">
+      <ul class="navbar-nav mr-auto">
+        <li class="nav-item">
+          <form class="form-inline my-2 my-lg-0">
+            <input class="form-control mr-sm-2" type="search" placeholder="Search">
+            <!-- <button type="button" class="btn btn-outline-secondary">Search</button>         -->
+          </form>
+        </li>
+        <li class="nav-item">
+          <button type="button" class="btn btn-primary btn" data-toggle="modal" data-target="#new_question">Ask a Question</button>
+        </li>
+      </ul>
+    </div>
+    <ul class="list-inline nav-list mb-0">   
+     <%# if user_signed_in? %>
+      <li class=" main_nav nav-item padding-3">
+        <%= link_to 'Profile', "#", class: "nav-link" %>
+
+      </li>
+    </ul>
+</nav>
