+++ app/views/questions/index.html.erb
@@ -1,0 +1,9 @@
+
+    <%= render 'banner' %>
+		<%= render 'navbar' %>
+    
+      
+    <div class="content" style="height: 2000px; background-color: #F5F5F6">
+    </div>
+
+
