--- db/migrate/20170628194008_create_answers.rb
@@ -1,10 +1,0 @@
-class CreateAnswers < ActiveRecord::Migration
-  def change
-    create_table :answers do |t|
-    	t.text :response
-    	t.references :user, index: true, foreign_key: true
-    	t.references :question, foreign_key: true
-    	t.timestamps null: false
-    end
-  end
-end
