+++ db/migrate/20170713005629_create_answers.rb
@@ -1,0 +1,8 @@
+class CreateAnswers < ActiveRecord::Migration
+  def change
+    create_table :answers do |t|
+
+      t.timestamps null: false
+    end
+  end
+end
